Update associatedLegendre.jl
This PR is the file entitled associatedLegendre.jl.  The first part with all the comments is like a readme.  The rest is the source code and numerous tests(requested) and exploration of area near |x| ~ 1.  As requested @btime from BenchmarkTools as well as @time gave similar results in that ZZLegendreP2 is light years ahead of scipy.special.lpmv.  Accuracy of ZZLegendreP2 (|z|=<1.) against lpmv and ZZLegendreP3 (z>1.) against Fortran is  good.