Pure Julia implementation of OpenLibm's `lgamma`, `lgamma_r`
Why bother with this port? The short answer is that for `Float64`,
there are considerable performance gains to be had. (See
https://github.com/andrewjradcliffe/OpenLibmPorts.jl for plots
produced on my machine -- I encourage you to reproduce them for yourself).

The long answer can be found in the linked package. To summarize for
those unfamiliar with large probabilistic models, log-posterior
computations often involve many unavoidable (i.e. cannot be eliminated
by using the unnormalized density) special function calls; in
particular, log-likelihoods involving the (log)gamma function are not
uncommon, e.g. negative binomial, beta-binomial. Given such a
likelihood function, each term in the log-likelihood sum would in
necessitate a loggamma call, hence, any reduction in the latency in
this particular special function translates can have a substantial
impact.

It is also a convenience to have a loggamma function which is
differentiable using `Enzyme` (though, there are some gaps in the
derivative -- at exactly 0.0, 1.0 and 2.0 -- but that is just a quirk
of applying AD to OpenLibm's loggamma implementation).

The author realizes that many things depend on `loggamma`, and my
intention is not to cause headaches for others. Fortunately, by
directly porting the OpenLibm implementation to Julia, we
achieve the same approximation, albeit, with slightly different
rounding due, presumably, to the difference between the
implementations of `log` in Julia and OpenLibm.